+++ PhotoOverlay/Presentation/PhotoList/View/PhotosViewController.swift
@@ -201,28 +201,3 @@ extension PhotosViewController {
         dataSource?.apply(snapShot, animatingDifferences: false)
     }
 }
-
-// MARK: - UICollectionView Extension
-
-private extension UICollectionView {
-    func dequeueReusableCell<T: UICollectionViewCell>(
-        withClass: T.Type,
-        indextPath: IndexPath
-    ) -> T {
-        guard let cell = self.dequeueReusableCell(
-            withReuseIdentifier: String(describing: T.self),
-            for: indextPath
-        ) as? T else {
-            return T()
-        }
-        
-        return cell
-    }
-    
-    func registerCell<T: UICollectionViewCell>(withClass: T.Type) {
-        self.register(
-            T.self,
-            forCellWithReuseIdentifier: String(describing: T.self)
-        )
-    }
-}
